Refactor: Moving route methods from server.go to app.go
