+++ server/server.go
@@ -1,11 +1,9 @@
 package server
 
 import (
-	"fmt"
 	"github.com/labstack/echo/v4"
 	"github.com/oscarmlage/wikingo/model"
 	"log"
-	"net/http"
 )
 
 // Depending on config we should open one store or other (Gorm, File,
@@ -32,25 +30,4 @@ func Serve() {
 
 	// Logger
 	e.Logger.Fatal(e.Start(":2323"))
-}
-
-func WikiHome(c echo.Context) error {
-	res := store.GetPage()
-	fmt.Println(res)
-	return c.String(http.StatusOK, "WikiHome")
-}
-
-func WikiAbout(c echo.Context) error {
-	id := c.Param("id")
-	return c.String(http.StatusOK, "About the wiki. id:"+id)
-}
-
-func WikiListPages(c echo.Context) error {
-	fmt.Println("WikiListPages")
-	res, err := store.GetAllPages()
-	if err != nil {
-		log.Panicln(err)
-	}
-	fmt.Println(res)
-	return c.String(http.StatusOK, "WikiListPages")
 }
